# execinfra: remove possible logging for each output row of processors