@yuzefovich 2022-10-27
Good point. I don't recall ever needing this kind of logging, and I don't think anyone ever used it. It doesn't seem particularly useful since it's extremely verbose because it would print out a log message for each row that passes through each processor of each query (i.e. including the intermediate steps), so I think we should just remove it.

TFTR!

bors r+